moved the test file (because functions it tests are now in the service layer)
